% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoquet.R
\name{apply_builtin_schema_registry_policies}
\alias{apply_builtin_schema_registry_policies}
\title{Applies schema policies that cannot be overridden}
\usage{
apply_builtin_schema_registry_policies(reg, profile = c("generic", "hcup"))
}
\description{
Applies schema policies that cannot be overridden
}
\keyword{internal}
