% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoquet.R
\docType{data}
\name{.DELIMITED_MEMORY_SAFETY_MULTIPLIER}
\alias{.DELIMITED_MEMORY_SAFETY_MULTIPLIER}
\title{fread, UTF-8 normalization, partition routing, and Arrow can coexist briefly.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.DELIMITED_MEMORY_SAFETY_MULTIPLIER
}
\description{
fread, UTF-8 normalization, partition routing, and Arrow can coexist briefly.
}
\keyword{datasets}
