% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoquet.R
\name{apply_builtin_schema_registry_policies}
\alias{apply_builtin_schema_registry_policies}
\title{Apply non-negotiable built-in schema policies}
\usage{
apply_builtin_schema_registry_policies(reg, profile = c("generic", "hcup"))
}
\description{
Apply non-negotiable built-in schema policies
}
\keyword{internal}
