% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoquet.R
\name{remove_parquet_manifest_rows}
\alias{remove_parquet_manifest_rows}
\title{Remove manifest rows matching one or more fields}
\usage{
remove_parquet_manifest_rows(
  ManifestPath,
  Database = NULL,
  TableName = NULL,
  SourcePath = NULL,
  ParquetPath = NULL
)
}
\description{
Remove manifest rows matching one or more fields
}
\keyword{internal}
