% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoquet.R
\name{write_parquet_manifest_atomic}
\alias{write_parquet_manifest_atomic}
\title{Write a manifest table through a temporary file replacement}
\usage{
write_parquet_manifest_atomic(manifest, ManifestPath)
}
\description{
Write a manifest table through a temporary file replacement
}
\keyword{internal}
