% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoquet.R
\name{write_schema_registry}
\alias{write_schema_registry}
\title{Write schema registry as an Excel workbook or CSV fallback}
\usage{
write_schema_registry(reg, SchemaRegistryPath)
}
\description{
Write schema registry as an Excel workbook or CSV fallback
}
\keyword{internal}
